[FEATURE] introduce json encoder, refactor json decor
