+++ Classes/Json/JsonEncoderInterface.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+/***
+ *
+ * This file is part of the "headless" Extension for TYPO3 CMS.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.md file that was distributed with this source code.
+ *
+ *  (c) 2020
+ *
+ ***/
+
+namespace FriendsOfTYPO3\Headless\Json;
+
+interface JsonEncoderInterface
+{
+    public function encode(array $data, int $encodeOptions = 0): string;
+}
